+++ TASKS.md
@@ -814,4 +814,4 @@
     B: TrainingSessionStateControl — AsyncSequence end-flow instead of polling 
     C: SessionFeature/Watch — HK channel for .workoutEnded in Watch-primary mode
     D: LiveActivity — Liquid Glass on Activity Lock Screen card
-    E: CleanUp
+    E: CleanUpx
